Refactor rendering into module
