+++ Cargo.toml
@@ -29,6 +29,9 @@ git = "https://github.com/gfx-rs/gfx-rs"
 git = "https://github.com/gfx-rs/gfx_macros"
 plugin = true
 
+[dependencies.cgmath]
+git = "https://github.com/bjz/cgmath-rs"
+
 [dev-dependencies.glutin]
 git = "https://github.com/tomaka/glutin"
